feat: client iter
